cryptarchia: move mk_block to test_common
