+++ cryptarchia/test_common.py
@@ -10,3 +10,19 @@ def mk_config() -> Config:
         epoch_period_nonce_stabilization=4,
         time=TimeConfig(slot_duration=1, chain_start_time=0),
     )
+
+
+def mk_block(
+    parent: Id, slot: int, coin: Coin, content=bytes(32), orphaned_proofs=[]
+) -> BlockHeader:
+    assert len(parent) == 32
+    from hashlib import sha256
+
+    return BlockHeader(
+        slot=Slot(slot),
+        parent=parent,
+        content_size=len(content),
+        content_id=sha256(content).digest(),
+        leader_proof=MockLeaderProof.new(coin, Slot(slot), parent=parent),
+        orphaned_proofs=orphaned_proofs,
+    )
